Improving base64
Just splitting it into .cpp and .h
